+++ plugin/test/net/groboclown/idea/p4ic/config/MockP4ConfigProject.java
@@ -210,6 +210,12 @@ public class MockP4ConfigProject extends P4ConfigProject {
         public boolean isPasswordStoredLocally() {
             return false;
         }
+
+        @Nullable
+        @Override
+        public String getClientHostname() {
+            return null;
+        }
     }